refactor log component
