+++ lib/public/views/Logs/Details/LogTreeViewModel.js
@@ -65,6 +65,7 @@ export class LogTreeViewModel extends Observable {
      * @returns {undefined}
      */
     collapseLog(id) {
+        console.log(id)
         const index = this._detailedPostsIds.indexOf(id, 0);
         if (index > -1) {
             this._detailedPostsIds.splice(index, 1);
