+++ lib/public/views/Logs/Details/logTreeViewComponent.js
@@ -14,8 +14,7 @@
 import { h } from '/js/src/index.js';
 import errorAlert from '../../../components/common/errorAlert.js';
 import spinner from '../../../components/common/spinner.js';
-import { logCollapsed } from '../../../components/Log/logCollapsed.js';
-import { logExpanded } from '../../../components/Log/logExpanded.js';
+import { logComponent } from '../../../components/Log/log.js'
 import scrollTo from '../../../utilities/scrollTo.js';
 import { frontLink } from '../../../components/common/navigation/frontLink.js';
 
@@ -41,22 +40,15 @@ export const logTreeViewComponent = (treeViewModel, authenticationToken) => {
                         'margin-left': `${2 * level}rem`,
                     },
                 },
-                isCollapsed
-                    ? logCollapsed(
-                        log,
-                        log.id === selectedLogId,
-                        !(log.title === parentLogTitle),
-                        () => treeViewModel.expandLog(log.id),
-                        () => treeViewModel.onCopyUrlSuccess(),
-                    )
-                    : logExpanded(
-                        log,
-                        log.id === selectedLogId,
-                        !(log.title === parentLogTitle),
-                        () => treeViewModel.collapseLog(log.id),
-                        () => treeViewModel.onCopyUrlSuccess(),
-                        authenticationToken
-                    ),
+                logComponent(
+                    log,
+                    log.id === selectedLogId,
+                    !(log.title === parentLogTitle),
+                    isCollapsed,
+                    () => treeViewModel.collapseLog(log.id),
+                    () => treeViewModel.expandLog(log.id),
+                    () => treeViewModel.onCopyUrlSuccess(),
+                )
             );
 
             return h('', [
